Pass along extra spacing for dataset selector
